refactored loginModal
